+++ src/components/ControlledSwitch/switch.scss
