+++ src/containers/Application.js
@@ -15,16 +15,13 @@ import Footer from './Footer';
 import Sidebar from './Sidebar';
 import SplashScreen from './SplashScreen';
 import Breadcrumbs from './Breadcrumbs';
-import { Spinner } from './Spinner';
 import ErrorHandler from './ErrorHandler';
+import Alerts from './Alerts';
+import { Spinner } from '../components/Spinner';
 
-import AlertsComponent from './AlertsComponent';
-import alertsReducer from '../reducers/alertReducer';
-
-import helpButtonReducer from '../reducers/helpButtonReducer';
-import sidebarReducer from '../reducers/sidebarReducer';
-
-import advancedModeReducer from '../reducers/advancedModeReducer';
+import alertsReducer from './Alerts/reducer';
+import sidebarReducer from './Sidebar/reducer';
+import headerHelpButtonReducer from './Header/reducer';
 
 import ReduxService from '../services/ReduxService';
 import ConfigService from '../config/ConfigService';
@@ -97,7 +94,7 @@ class Application extends Component {
 
 		this.ReduxService.addReducer("alerts", alertsReducer);
 		this.ReduxService.addReducer("advmode", advancedModeReducer);
-		this.ReduxService.addReducer("helpButton", helpButtonReducer);
+		this.ReduxService.addReducer("helpButton", headerHelpButtonReducer);
 		this.ReduxService.addReducer("sidebar", sidebarReducer);
 
 		this.DefaultPath = props.defaultpath;
@@ -480,7 +477,7 @@ class Application extends Component {
 			<Provider store={this.Store}>
 				<div className="app">
 					<Suspense fallback={<></>}>
-						<AlertsComponent app={this} />
+						<Alerts app={this} />
 						<TenantSelectionCard app={this} />
 					</Suspense>
 					<SplashScreen app={this} />
@@ -500,7 +497,7 @@ class Application extends Component {
 					<Fade in={this.state.networking > 0} timeout={50} >
 						<div className="networking-indicator progress-bar progress-bar-animated progress-bar-striped" ></div>
 					</Fade>
-					<AlertsComponent app={this} />
+					<Alerts app={this} />
 					{this.Config.get('title') != null && this.Config.get('title') != undefined ?
 						<Helmet>
 							<title>{this.Config.get('site_title') ? this.Config.get('site_title') + " | " + this.Config.get('title') : this.Config.get('title')}</title>
@@ -607,3 +604,24 @@ class Navigation {
 
 
 export default withRouter(Application);
+
+
+// Advanced mode
+
+const advModeInitialState = {
+	enabled: false,
+}
+
+function advancedModeReducer(state = advModeInitialState, action) {
+	switch (action.type) {
+
+		case SET_ADVANCED_MODE: {
+			return Object.assign({}, state, {
+				enabled: action.enabled
+			})
+		}
+
+		default:
+			return state
+	}
+}
