+++ src/index.js
@@ -3,17 +3,17 @@ import "babel-polyfill";
 export { default as Application } from './containers/Application';
 export { getBaseName } from './basename'; // THIS FUNCTION HAS BEEN DEPRECATED - DELETE AFTER SEP 2021
 
-export { DateTime } from './containers/DateTime';
-export { Humanize } from './containers/Humanize';
-export { DataTable } from './containers/DataTable';
+export { DateTime } from './components/DateTime';
+export { Humanize } from './components/Humanize';
+export { DataTable } from './components/DataTable';
 export { ButtonWithAuthz } from './modules/auth/ButtonWithAuthz';
-export { Spinner } from './containers/Spinner';
-export { KnowledgeBase } from './containers/KnowledgeBase';
+export { Spinner } from './components/Spinner';
+export { KnowledgeBase } from './components/KnowledgeBase';
 export { default as SplashScreen } from './containers/SplashScreen';
-export { default as Pagination } from './containers/Pagination';
-export { default as ControlledSwitch } from './containers/ControlledSwitch';
-export { default as UncontrolledSwitch } from './containers/UncontrolledSwitch';
-export { Credentials } from './containers/Credentials';
-export { CellContentLoader, ChartLoader } from './containers/ContentLoader';
+export { default as Pagination } from './components/Pagination';
+export { default as ControlledSwitch } from './components/ControlledSwitch';
+export { default as UncontrolledSwitch } from './components/UncontrolledSwitch';
+export { Credentials } from './components/Credentials';
+export { CellContentLoader, ChartLoader } from './components/ContentLoader';
 
 import "./style.scss";
